==> 07_v3_GenusPlots.Rmd <==
---
title: "07_GenusPlots"
output: html_document
date: "2025-04-28"
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(readxl)
library(dplyr)
library(tidyverse)
library(funspace)
library(missForest)
library(caret)
library(RColorBrewer)
library(ggbiplot)
library(viridis)
library(forcats) 
library(patchwork)
library(lme4)
library(ggrepel)
library(dplyr)
library(broom)
library(purrr)
library(stringr)
```

```{r}
imputed <- read_csv(file = "imputed_fecundity_v3.csv")

imputed_long <- imputed %>%
  dplyr::select(Species, Genus, family_molecules, GrowthForm_CTD, fecundity, Colony_fecundity, IndoPacific_fecundity) %>%
  pivot_longer(cols = c(fecundity, Colony_fecundity, IndoPacific_fecundity),
               names_to = "Fecundity_Type", values_to = "Fecundity_Value")

imputed_long <- imputed_long%>%
  mutate(family_molecules = case_when(
    Genus == "Caulastraea" ~ "Merulinidae",
    TRUE ~ family_molecules
  ))

imputed_long <- imputed_long %>%
  mutate(family_molecules = case_when(
    Species == "Acropora muricata" ~ "Acroporidae",
    Species == "Favites micropentagonus" ~ "Merulinidae",
    TRUE ~ family_molecules  # Keep existing values for all other rows
  ))

fecundity_data <- imputed_long %>%
  filter(Fecundity_Type == "fecundity")

colony_data <- imputed_long %>%
  filter(Fecundity_Type == "Colony_fecundity")

indo_pacific_data <- imputed_long %>%
  filter(Fecundity_Type == "IndoPacific_fecundity")
```


```{r}

# Step 1: Identify the top 10 genera for each fecundity type
top10_fecundity <- imputed %>%
  group_by(Genus) %>%
  summarize(Total_Fecundity = sum(fecundity, na.rm = TRUE)) %>%
  arrange(desc(Total_Fecundity))%>%
  slice(1:6) %>%
  pull(Genus)

top10_colony <- imputed %>%
  group_by(Genus) %>%
  summarize(Total_Colony_Fecundity = sum(Colony_fecundity, na.rm = TRUE)) %>%
  arrange(desc(Total_Colony_Fecundity)) %>%
  slice(1:6) %>%
  pull(Genus)

top10_indo_pacific <- imputed %>%
  group_by(Genus) %>%
  summarize(Total_IndoPacific_Fecundity = sum(IndoPacific_fecundity, na.rm = TRUE)) %>%
  arrange(desc(Total_IndoPacific_Fecundity)) %>%
  slice(1:6) %>%
  pull(Genus)

# Step 2: Create separate dataframes for each fecundity type, categorizing genera as "Top 10" or "Other"
fecundity_data_genus <- imputed %>%
  mutate(Genus_Category = ifelse(Genus %in% top10_fecundity, Genus, "Other")) %>%
  group_by(Genus_Category) %>%
  summarize(Value = sum(fecundity, na.rm = TRUE)) %>%
  mutate(Fecundity_Type = "Fecundity")

colony_data_genus <- imputed %>%
  mutate(Genus_Category = ifelse(Genus %in% top10_colony, Genus, "Other")) %>%
  group_by(Genus_Category) %>%
  summarize(Value = sum(Colony_fecundity, na.rm = TRUE)) %>%
  mutate(Fecundity_Type = "Colony Fecundity")

indo_pacific_data_genus <- imputed %>%
  mutate(Genus_Category = ifelse(Genus %in% top10_indo_pacific, Genus, "Other")) %>%
  group_by(Genus_Category) %>%
  summarize(Value = sum(IndoPacific_fecundity, na.rm = TRUE)) %>%
  mutate(Fecundity_Type = "IndoPacific Fecundity")


genus_color_scheme <- c(
  "Acropora" = "#9E0142",
  "Favites" = "#3288BD",
  "Goniopora" = "#5E4FA2",
  "Isopora" = "#ABDDA4",
  "Leptastrea" = "lightblue",
  "Montipora" = "#F46D43",
  "Other" = "grey",
  "Pavona" = "#66C2A5",
  "Ctenactis" = "turquoise3",
  "Porites" = "#E6F598",
  "Echinopora" = "pink"
  
)

##########Fecundity 

fecundity_data_genus <- fecundity_data_genus%>%
  mutate(Proportion = Value / sum(Value))%>%
  arrange(Proportion)

fecundity_genus_order <- fecundity_data_genus %>%
  group_by(Genus_Category) %>%
  summarise(Total_Proportion = sum(Proportion)) %>%
  arrange(Total_Proportion)  # Change to descending order

# Apply ordering to the factor
fecundity_data_genus <- fecundity_data_genus %>%
  mutate(Genus_Category = factor(Genus_Category, levels = fecundity_genus_order$Genus_Category))

g1 <- ggplot(fecundity_data_genus, aes(x = Fecundity_Type, y = Proportion, fill = Genus_Category)) +
  geom_bar(stat = "identity", position = "fill", color = "black") +
  labs(
    x = NULL,
    y = "Proportion of Total Reproductive Potential",
    fill = "Genus"
  ) +
  theme(
    legend.position = "none",  # Hide legend for this plot
    panel.grid = element_blank(),
    panel.background = element_rect(fill = "white"),
    plot.background = element_rect(fill = "white"),
    axis.line = element_line(color = "black"),
    axis.text.y = element_text(size = 14),
    axis.text.x = element_text(size = 14),  # Increase x-axis text size
    axis.title.x = element_text(size = 14, face = "bold"),
    legend.text = element_text(size = 14),  # Increase legend text size
    legend.title = element_text(size = 16)  # Increase legend title size
  ) +
  annotate("text", x = -Inf, y = Inf, label = "A", hjust = -0.1, vjust = 1.3, size = 8, fontface = "bold")+
  scale_y_continuous(labels = scales::percent) +
  scale_fill_manual(values = genus_color_scheme)+
  scale_x_discrete(labels = function(x) {  # Rename x-axis categories
    ifelse(
      x == "Fecundity", expression("Fecundity per" ~ cm^2),
      ifelse(
        x == "Colony Fecundity", expression("Colony Fecundity"),
        ifelse(x == "IndoPacific Fecundity", expression("Total Fecundity"), x)
      )
    )
  })


colony_data_genus <- colony_data_genus%>%
  mutate(Proportion = Value / sum(Value))%>%
  arrange(Proportion)

colony_genus_order <- colony_data_genus %>%
  group_by(Genus_Category) %>%
  summarise(Total_Proportion = sum(Proportion)) %>%
  arrange(Total_Proportion)  # Change to descending order

# Apply ordering to the factor
colony_data_genus <- colony_data_genus %>%
  mutate(Genus_Category = factor(Genus_Category, levels = colony_genus_order$Genus_Category))

g2 <- ggplot(colony_data_genus, aes(x = Fecundity_Type, y = Proportion, fill = Genus_Category)) +
  geom_bar(stat = "identity", position = "fill", color = "black") +
  labs(
    x = NULL,
    y = "Proportion of Total Reproductive Potential",
    fill = "Genus"
  ) +
  theme(
    legend.position = "none",  # Hide legend for this plot
    panel.grid = element_blank(),
    panel.background = element_rect(fill = "white"),
    plot.background = element_rect(fill = "white"),
    axis.line = element_line(color = "black"),
    axis.title.y = element_blank(),
    axis.text.y = element_blank(),
    axis.ticks.y = element_blank(),
    axis.line.y = element_blank(),
    axis.text.x = element_text(size = 14),  # Increase x-axis text size
    axis.title.x = element_text(size = 14, face = "bold"),
    legend.text = element_text(size = 14),  # Increase legend text size
    legend.title = element_text(size = 16)  # Increase legend title size
  ) +
  scale_y_continuous(labels = scales::percent) +
  scale_fill_manual(values = genus_color_scheme)+
  scale_x_discrete(labels = function(x) {  # Rename x-axis categories
    ifelse(
      x == "Fecundity", expression("Fecundity per" ~ cm^2),
      ifelse(
        x == "Colony Fecundity", expression("Colony Fecundity"),
        ifelse(x == "IndoPacific Fecundity", expression("Total Fecundity"), x)
      )
    )
  })


indo_pacific_data_genus <- indo_pacific_data_genus %>%
  mutate(Proportion = Value / sum(Value))%>%
  arrange(Proportion)

indo_pacific_genus_order <- indo_pacific_data_genus %>%
  group_by(Genus_Category) %>%
  summarise(Total_Proportion = sum(Proportion)) %>%
  arrange(Total_Proportion)  # Change to descending order

# Apply ordering to the factor
indo_pacific_data_genus <- indo_pacific_data_genus %>%
  mutate(Genus_Category = factor(Genus_Category, levels = indo_pacific_genus_order$Genus_Category))

g3 <- ggplot(indo_pacific_data_genus, aes(x = Fecundity_Type, y = Proportion, fill = Genus_Category)) +
  geom_bar(stat = "identity", position = "fill", color = "black") +
  labs(
    x = NULL,
    y = "Proportion of Total Reproductive Potential",
    fill = "Genus"
  ) +
  theme(
    legend.position = "none",  # Hide legend for this plot
    panel.grid = element_blank(),
    panel.background = element_rect(fill = "white"),
    plot.background = element_rect(fill = "white"),
    axis.line = element_line(color = "black"),
    axis.title.y = element_blank(),
    axis.text.y = element_blank(),
    axis.ticks.y = element_blank(),
    axis.line.y = element_blank(),
    axis.text.x = element_text(size = 14),  # Increase x-axis text size
    axis.title.x = element_text(size = 14, face = "bold"),
    legend.text = element_text(size = 14),  # Increase legend text size
    legend.title = element_text(size = 16)  # Increase legend title size
  ) +
  scale_y_continuous(labels = scales::percent) +
  scale_fill_manual(values = genus_color_scheme)+
  scale_x_discrete(labels = function(x) {  # Rename x-axis categories
    ifelse(
      x == "Fecundity", expression("Fecundity per" ~ cm^2),
      ifelse(
        x == "Colony Fecundity", expression("Colony Fecundity"),
        ifelse(x == "IndoPacific Fecundity", expression("Total Fecundity"), x)
      )
    )
  })

# Create a dummy dataframe for the legend
legend_data <- data.frame(Genus = names(genus_color_scheme), Value = 1)

# Create a dummy plot that only generates the legend
legend_plot <- ggplot(data.frame(Genus = names(genus_color_scheme)), aes(fill = Genus)) +
  geom_point(aes(x = 1, y = 1), shape = 22, size = 4) +  # Invisible points to create a legend
  scale_fill_manual(values = genus_color_scheme) +  # Use predefined color scheme
  theme_void() +  # Remove all axes and background
  theme(legend.position = "top",  # Place legend at top
        legend.title = element_text(size = 14, face = "bold"),
        legend.text = element_text(size = 12)) +
  labs(fill = "Genus")  # Legend title

genus_final_plot <- (legend_plot / (g1 | g2 | g3)) +  # Legend on top, plots side by side
  plot_layout(heights = c(0.15, 1))  # Adjust height ratio to make legend smaller

# Display the final plot
genus_final_plot

ggsave(filename = "../methods/plots/genus_fecundity_v3.pdf", 
       plot = genus_final_plot, 
       width = 15, height = 10, units = "in")

```

```{r}

# Automatically abbreviate species names
fecundity_data <- fecundity_data %>%
  mutate(Species = str_replace(Species, 
                               "^([A-Za-z])[a-z]+\\s",  # Capture the first letter of the genus
                               "\\1. ")) 

fecundity_data <- fecundity_data%>%
  group_by(Genus)%>%
  mutate(Genus_Mean = mean(Fecundity_Value))%>%
  mutate(Difference = Fecundity_Value - Genus_Mean)%>%
  ungroup()
  

extreme_species <- fecundity_data %>%
  group_by(Genus) %>%
  filter(Fecundity_Value == max(Fecundity_Value) | Fecundity_Value == min(Fecundity_Value)) %>%
  ungroup()


all <- ggplot(fecundity_data, aes(x = Genus_Mean, y = Fecundity_Value, color = Difference > 0)) +
  geom_point(size = 3, alpha = 0.7) +  # Each point is a species
  geom_point(aes(x = Genus_Mean, y = Genus_Mean, shape = "Genus Mean Value"),  
             size = 4, color = "grey", fill = "grey") +  # Mean point
  geom_abline(slope = 1, intercept = 0, linetype = "dashed", color = "black") +  # 1:1 reference line
  geom_text_repel(
  data = extreme_species, aes(label = Species),
  size = 4, color = "black",
  box.padding = 1, point.padding = 0.5,
  segment.color = "black", segment.size = 0.5,
  force = 5, direction = "y")+
  labs(
    x = expression("Genus mean fecundity (eggs per " *cm^2* ")"),
    y = expression("Species fecundity (eggs per " *cm^2* ")"),
    color = NULL,
    shape = NULL  # Remove shape legend title
  ) +
  theme_minimal() +
  theme(
    legend.position = c(0.95, 0.1),  # Move legend to bottom-right area
    legend.justification = c(1, 0),  # Align legend to bottom-right
    panel.grid = element_blank(),
    axis.line = element_line(color = "black"),  # Add axis lines
    axis.text = element_text(size = 12),  # Adjust axis text size,
    axis.title = element_text(size = 14, face = "bold"),  # Make axis titles bold
    strip.text = element_text(size = 14, face = "bold")
  ) +
  scale_color_manual(
    values = c("TRUE" = "lightgreen", "FALSE" = "red"),
    labels = c("FALSE" = "Less than Genus Mean", "TRUE" = "Greater than Genus Mean")  # Update labels
  ) +
  scale_shape_manual(values = c("Genus Mean Value" = 23))  # Assign square shape for mean in legend

ggsave(filename = "../methods/plots/GenusSpecies_fecundity_all_v2.pdf", 
       plot = all, 
       width = 20, height = 10, units = "in")
all
```



```{r}
top_genus_diff <- fecundity_data %>%
  group_by(Genus) %>%
  summarise(
    max_diff = max(Difference, na.rm = TRUE),
    min_diff = min(Difference, na.rm = TRUE),
    total_difference = max_diff - min_diff
  ) %>%
  arrange(desc(total_difference)) %>%
  slice_head(n = 8)

top10diff_genus_fecundity <- fecundity_data %>%
  filter(Genus %in% top_genus_diff$Genus)

extreme_species <- top10diff_genus_fecundity %>%
  group_by(Genus) %>%
  filter(Fecundity_Value == max(Fecundity_Value) | Fecundity_Value == min(Fecundity_Value)) %>%
  ungroup()


top10 <- ggplot(top10diff_genus_fecundity, aes(x = Genus_Mean, y = Fecundity_Value, color = Difference > 0)) +
  geom_point(size = 3, alpha = 0.7) +  # Each point is a species
  geom_point(aes(x = Genus_Mean, y = Genus_Mean, shape = "Genus Mean Value"),  
             size = 4, color = "grey", fill = "grey") +  # Mean point
  geom_abline(slope = 1, intercept = 0, linetype = "dashed", color = "black") +  # 1:1 reference line
  geom_text_repel(
  data = extreme_species, aes(label = Species),
  size = 4, color = "black",
  box.padding = 1, point.padding = 0.5,
  segment.color = "black", segment.size = 0.5,
  force = 5, direction = "y")+
  labs(
    x = expression("Genus mean fecundity (eggs per " *cm^2* ")"),
    y = expression("Species fecundity (eggs per " *cm^2* ")"),
    color = NULL,
    shape = NULL  # Remove shape legend title
  ) +
  theme_bw() +  # White background
  theme(
    legend.position = c(0.95, 0.1),  # Move legend to bottom-right area
    legend.justification = c(1, 0),  # Align legend to bottom-right
    panel.grid = element_blank(),               # No grid lines
    panel.border = element_rect(color = "black"), # Add black border
    axis.line = element_line(color = "black"),
    axis.text = element_text(size = 12),
    axis.title = element_text(size = 14, face = "bold"),
    strip.text = element_text(size = 14, face = "bold")
  )+
  scale_color_manual(
    values = c("TRUE" = "lightgreen", "FALSE" = "red"),
    labels = c("FALSE" = "Less than Genus Mean", "TRUE" = "Greater than Genus Mean")  # Update labels
  )+
  facet_wrap(~Genus)

ggsave(filename = "../methods/plots/GenusSpecies_fecundity_top10_v2.pdf", 
       plot = top10, 
       width = 20, height = 10, units = "in")

top10
```
```{r}
top_genus_diff <- fecundity_data %>%
  group_by(Genus) %>%
  summarise(
    max_diff = max(Difference, na.rm = TRUE),
    min_diff = min(Difference, na.rm = TRUE),
    total_difference = max_diff - min_diff
  ) %>%
  arrange(desc(total_difference)) %>%
  slice_head(n = 8)

top10diff_genus_fecundity <- fecundity_data %>%
  filter(Genus %in% top_genus_diff$Genus)

extreme_species <- top10diff_genus_fecundity %>%
  group_by(Genus) %>%
  filter(Fecundity_Value == max(Fecundity_Value) | Fecundity_Value == min(Fecundity_Value)) %>%
  ungroup()


top10 <- ggplot(top10diff_genus_fecundity, aes(x = Genus_Mean, y = Fecundity_Value, color = Difference > 0)) +
  geom_point(size = 3, alpha = 0.7) +  # Each point is a species
  geom_point(aes(x = Genus_Mean, y = Genus_Mean, shape = "Genus Mean Value"),  
             size = 4, color = "grey", fill = "grey") +  # Mean point
  geom_abline(slope = 1, intercept = 0, linetype = "dashed", color = "black") +  # 1:1 reference line
  geom_text_repel(
  data = extreme_species, aes(label = Species),
  size = 4, color = "black",
  box.padding = 1, point.padding = 0.5,
  segment.color = "black", segment.size = 0.5,
  force = 5, direction = "y")+
  labs(
    x = expression("Genus mean fecundity (eggs per " *cm^2* ")"),
    y = expression("Species fecundity (eggs per " *cm^2* ")"),
    color = NULL,
    shape = NULL  # Remove shape legend title
  ) +
  theme_bw() +  # White background
  theme(
    legend.position = c(0.95, 0.1),  # Move legend to bottom-right area
    legend.justification = c(1, 0),  # Align legend to bottom-right
    panel.grid = element_blank(),               # No grid lines
    panel.border = element_rect(color = "black"), # Add black border
    axis.text.x = element_blank(),   # Removes tick labels
    axis.ticks.x = element_blank(), 
    axis.line = element_line(color = "black"),
    axis.text = element_text(size = 12),
    axis.title = element_text(size = 14, face = "bold"),
    strip.text = element_text(size = 14, face = "bold")
  )+
  scale_color_manual(
    values = c("TRUE" = "lightgreen", "FALSE" = "red"),
    labels = c("FALSE" = "Less than Genus Mean", "TRUE" = "Greater than Genus Mean")  # Update labels
  )+
  scale_x_continuous(breaks = scales::pretty_breaks(n = 0))+
  facet_wrap(~Genus, scales = "free")


top10

ggsave(filename = "../methods/plots/GenusSpecies_fecundity_top10_freescale_v3.pdf", 
       plot = top10, 
       width = 20, height = 10, units = "in")
```
```{r}


indo_pacific_proportions <- imputed %>%
  group_by(Genus)%>%
  mutate(indo_fecund_genus = sum(IndoPacific_fecundity))%>%
  select(Genus, indo_fecund_genus)%>%
  unique()%>%
  ungroup()%>%
  mutate(proportions = indo_fecund_genus / sum(indo_fecund_genus))
  
sd(indo_pacific_proportions$proportions) / sqrt(length(indo_pacific_proportions$proportions))

genus_indopacific <- indo_pacific_data_genus%>%
  arrange()%>%
  slice(1:6)

sum(genus_indopacific$Proportion)

sd(genus_indopacific$Proportion) / sqrt(length(genus_indopacific$Proportion))

```

```{r}

colony_proportions <- imputed %>%
  group_by(Genus)%>%
  mutate(colony_genus = sum(Colony_fecundity))%>%
  select(Genus, colony_genus)%>%
  unique()%>%
  ungroup()%>%
  mutate(proportions = colony_genus / sum(colony_genus))

sd(colony_proportions$proportions) / sqrt(length(colony_proportions$proportions))

genus_cology <- colony_data_genus%>%
  arrange()%>%
  slice(1:6)

sum(genus_cology$Proportion)

sd(genus_cology$Proportion) / sqrt(length(genus_cology$Proportion))

```

```{r}

fecundity_proportions <- imputed %>%
  group_by(Genus)%>%
  mutate(fecund_genus = sum(fecundity))%>%
  select(Genus, fecund_genus)%>%
  unique()%>%
  ungroup()%>%
  mutate(proportions = fecund_genus / sum(fecund_genus))

sd(fecundity_proportions$proportions) / sqrt(length(fecundity_proportions$proportions))

genus_fecundity <- fecundity_data_genus%>%
  arrange()%>%
  slice(1:6)

sum(genus_fecundity$Proportion)

sd(genus_fecundity$Proportion) / sqrt(length(genus_fecundity$Proportion))

```
